pridanie zobar nevyhnutnych tried nefunkcnost JdpdTemplate chyba nieco v pom.xml

+++ src/main/java/com/example/projektvolby/Kandidati.java
@@ -1,10 +1,10 @@
 package com.example.projektvolby;
 
 public class Kandidati {
-	private int id;
+	private long id;
     private String meno;
     private String priezvisko;
-    private int vek;
+    private long vek;
     private long pocet_hlasov;
     private Strany Strana;
